moved checkpointing funcs to utils
