+++ networkAlignmentAnalysis/experiments/experiment.py
@@ -225,35 +225,6 @@ class Experiment(ABC):
             cname = name + f"{idx}"
             save(net, self.get_network_path(cname))
 
-    def load_checkpoints(self, nets, optimizers, device):
-        """
-        Method for loading presaved checkpoint during training.
-        TODO: device handling for passing between gpu/cpu
-        """
-        checkpoint = load(self.get_checkpoint_path())
-        net_ids = sorted([key for key in checkpoint if key.startswith('model_state_dict')])
-        opt_ids = sorted([key for key in checkpoint if key.startswith('optimizer_state_dict')])
-        assert all([oi.split('_')[-1] == ni.split('_')[-1] for oi, ni in zip(opt_ids, net_ids)]), (
-            'nets and optimizers cannot be matched up from checkpoint'
-        )
-
-        # orig_device = checkpoint['device']
-        orig_device = 'cpu'
-        if (orig_device == 'cuda') & (device == 'cpu'):
-            [net.load_state_dict(checkpoint.pop(net_id), map_location=device)
-             for net, net_id in zip(nets, net_ids)]
-
-        [net.load_state_dict(checkpoint.pop(net_id))
-        for net, net_id in zip(nets, net_ids)]
-
-        if device == 'cuda':
-            [net.to(device) for net in nets]
-
-        [opt.load_state_dict(checkpoint.pop(opt_id))
-         for opt, opt_id in zip(optimizers, opt_ids)]
-
-        return nets, optimizers, checkpoint
-
     @abstractmethod
     def main(self) -> Tuple[Dict, List[TorchModule]]:
         """
