+++ networkAlignmentAnalysis/utils.py
@@ -1,12 +1,12 @@
 from typing import List
+# -- deprecation block --
+from warnings import warn
+
 import numpy as np
 import torch
 from matplotlib import pyplot as plt
 from torchvision import transforms
 
-
-# -- deprecation block --
-from warnings import warn
 
 def avg_align_by_layer(full):
     warn("avg_align_by_layer is deprecated, change to avg_value_by_layer!", DeprecationWarning, stacklevel=2)
@@ -393,3 +393,44 @@ def plot_rf(rf, width, alignment=None, alignBounds=None, showRFs=None, figSize=5
         ax.set_aspect('equal')
     fig.subplots_adjust(wspace=0.0, hspace=0.0)
     return fig
+
+
+def save_checkpoint(nets, optimizers, results, path):
+    """
+    Method for saving checkpoints for networks throughout training.
+    """
+    multi_model_ckpt = {f'model_state_dict_{i}': net.state_dict()
+                        for i, net in enumerate(nets)}
+    multi_optimizer_ckpt = {f'optimizer_state_dict_{i}': opt.state_dict()
+                            for i, opt in enumerate(optimizers)}
+    checkpoint = results | multi_model_ckpt | multi_optimizer_ckpt
+    torch.save(checkpoint, path)
+
+
+def load_checkpoints(nets, optimizers, device, path):
+    """
+    Method for loading presaved checkpoint during training.
+    TODO: device handling for passing between gpu/cpu
+    """
+    checkpoint = torch.load(path)
+    net_ids = sorted([key for key in checkpoint if key.startswith('model_state_dict')])
+    opt_ids = sorted([key for key in checkpoint if key.startswith('optimizer_state_dict')])
+    assert all([oi.split('_')[-1] == ni.split('_')[-1] for oi, ni in zip(opt_ids, net_ids)]), (
+        'nets and optimizers cannot be matched up from checkpoint'
+    )
+
+    orig_device = checkpoint['device']
+    if (orig_device == 'cuda') & (device == 'cpu'):
+        [net.load_state_dict(checkpoint.pop(net_id), map_location=device)
+            for net, net_id in zip(nets, net_ids)]
+
+    [net.load_state_dict(checkpoint.pop(net_id))
+        for net, net_id in zip(nets, net_ids)]
+
+    if device == 'cuda':
+        [net.to(device) for net in nets]
+
+    [opt.load_state_dict(checkpoint.pop(opt_id))
+        for opt, opt_id in zip(optimizers, opt_ids)]
+
+    return nets, optimizers, checkpoint
